修正頁面不會更新和更新 endpoint 後會無故刪除資料問題

+++ webservice-server/src/main/java/com/dynamicwebservice/service/impl/DynamicWebServiceImpl.java
@@ -162,7 +162,9 @@ public class DynamicWebServiceImpl implements DynamicWebService {
     public void updateWebService(EndpointDTO endpointDTO) throws FileNotFoundException {
         EndpointEntity endpointEntity = endpointRepository.findByUuId(endpointDTO.getId());
         saveWebService(endpointDTO);
-        endpointRepository.delete(endpointEntity);
+        if (!endpointDTO.getPublishUrl().equals(endpointEntity.getPublishUrl())) {
+            endpointRepository.delete(endpointEntity);
+        }
     }
 
     @Override
